dsm/requirements.txt
asgiref==3.6.0
astroid==2.5
backports.zoneinfo==0.2.1
CacheControl==0.12.11
cachetools==5.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.4
cryptography==39.0.0
Deprecated==1.2.13
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-location-field==2.1.0
environ==1.0
firebase-admin==6.0.1
gcloud==0.18.3
geographiclib==2.0
geopy==2.3.0
google-api-core==2.11.0
google-api-python-client==2.71.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.3
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
httplib2==0.21.0
idna==3.4
isort==5.7.0
jwcrypto==1.4.2
jws==0.1.3
lazy-object-proxy==1.5.2
mccabe==0.6.1
msgpack==1.0.4
oauth2client==4.1.3
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
PyJWT==2.6.0
pylint==2.7.0
pyparsing==3.0.9
Pyrebase4==4.6.0
python-decouple==3.7
python-jwt==2.0.1
requests==2.28.1
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
wincertstore==0.2
wrapt==1.14.1
